create strategies, add optimizedMinimumSearch into abstract class
